task_42 is ready

+++ Program.cs
@@ -3,13 +3,13 @@
 
 // b1 = 2, k1 = 5, b2 = 4, k2 = 9 -> (-0,5; -0,5)
 System.Console.WriteLine("Введите значение b1");
-int b1 = Convert.ToInt32(Console.ReadLine());
+double b1 = Convert.ToDouble(Console.ReadLine());
 System.Console.WriteLine("Введите значение k2");
-int  k1 = Convert.ToInt32(Console.ReadLine());
+double  k1 = Convert.ToDouble(Console.ReadLine());
 System.Console.WriteLine("Введите значение b2");
-int b2 = Convert.ToInt32(Console.ReadLine());
+double b2 = Convert.ToDouble(Console.ReadLine());
 System.Console.WriteLine("Введите значение k2");
-int k2 = Convert.ToInt32(Console.ReadLine());
+double k2 = Convert.ToDouble(Console.ReadLine());
 double x = (b2 - b1)/(k1 - k2);
 double y = k1 * x + b1;
 System.Console.WriteLine($"({x}; {y})");
